requirements.txt
Faker==37.1.0
protobuf==6.30.1
python-dotenv==1.0.1
